requirements.txt
# Audio Stream - Core dependencies
# Install with: pip install -r requirements.txt

# WebRTC audio streaming
aiortc>=1.5.0
aiohttp>=3.8.0
av>=9.0.0

# Audio capture
sounddevice>=0.4.5
numpy>=1.21.0

# GUI (CustomTkinter version)
customtkinter>=5.0.0

# QR code generation
qrcode[pil]>=7.3
pillow>=9.0.0
